move trie traits into trie module
